Split LLVM call and asm emission
