+++ crates/dialect-llvm/src/export/op_emission/casts.rs
@@ -6,13 +6,35 @@
 use std::collections::HashMap;
 use std::fmt::Write;
 
-use pliron::{op::Op, operation::Operation, r#type::Typed, value::Value};
+use pliron::{context::Ptr, op::Op, operation::Operation, r#type::Typed, value::Value};
 
 use crate::ops;
 
 use super::super::ModuleExportState;
 
 impl<'a> ModuleExportState<'a> {
+    pub(super) fn export_cast(
+        &self,
+        op_name: &str,
+        op: Ptr<Operation>,
+        value_names: &HashMap<Value, String>,
+        output: &mut String,
+    ) -> Result<(), String> {
+        let op_ref = op.deref(self.ctx);
+        let res = op_ref.get_result(0);
+        let val = op_ref.get_operand(0);
+        let res_name = value_names.get(&res).unwrap();
+
+        write!(output, "  {res_name} = {op_name} ").unwrap();
+        self.export_type(val.get_type(self.ctx), output)?;
+        write!(output, " ").unwrap();
+        self.export_value(val, value_names, output)?;
+        write!(output, " to ").unwrap();
+        self.export_type(res.get_type(self.ctx), output)?;
+        writeln!(output).unwrap();
+        Ok(())
+    }
+
     pub(super) fn export_zext_op(
         &self,
         op_ref: &Operation,
