+++ crates/dialect-llvm/src/export/op_emission/mod.rs
@@ -6,26 +6,17 @@
 use std::collections::HashMap;
 use std::fmt::Write;
 
-use pliron::{
-    basic_block::BasicBlock,
-    builtin::op_interfaces::{CallOpCallable, CallOpInterface},
-    context::Ptr,
-    op::Op,
-    operation::Operation,
-    r#type::Typed,
-    value::Value,
-};
+use pliron::{basic_block::BasicBlock, context::Ptr, op::Op, operation::Operation, value::Value};
 
-use crate::{
-    ops,
-    types::{FuncType, VoidType},
-};
+use crate::ops;
 
-use super::{ModuleExportState, strip_device_prefix};
+use super::ModuleExportState;
 
 mod aggregate;
 mod arithmetic;
+mod asm;
 mod atomic;
+mod calls;
 mod casts;
 mod memory;
 mod terminator;
@@ -183,125 +174,15 @@ impl<'a> ModuleExportState<'a> {
                 self.export_fcmp_op(&op_ref, fcmp, value_names, output)?;
             }
 
-            // --- Calls ---
-            // LLVM call instruction format:
-            //   - Non-void: %result = call <ret_type> @func(<args>)
-            //   - Void:     call void @func(<args>)
-            //
-            // IMPORTANT: Void-returning calls must NOT have a result assignment.
-            // Invalid: "%v1 = call void @foo()" - llc will reject this!
-            // Valid:   "call void @foo()"
             id if id == ops::CallOp::get_opid_static() => {
                 let call = op_obj.as_ref().downcast_ref::<ops::CallOp>().unwrap();
-                let callee = call.callee(self.ctx);
-
-                // Extract return type from the call's function type to determine
-                // if this is a void call (no result assignment) or value call
-                let func_ty = call.callee_type(self.ctx);
-                let func_ty_ref = func_ty.deref(self.ctx);
-                let llvm_func_ty = func_ty_ref.downcast_ref::<FuncType>().unwrap();
-                let ret_ty = llvm_func_ty.result_type();
-                let is_void = ret_ty.deref(self.ctx).is::<VoidType>();
-
-                // Void calls: "call void @func(...)"
-                // Non-void:   "%vN = call <type> @func(...)"
-                if is_void {
-                    write!(output, "  call void").unwrap();
-                } else {
-                    let res = op_ref.get_result(0);
-                    let res_name = value_names.get(&res).unwrap();
-                    write!(output, "  {res_name} = call ").unwrap();
-                    self.export_type(ret_ty, output)?;
-                }
-
-                // Track if callee is a convergent intrinsic
-                let mut is_convergent_call = false;
-
-                // Callee can be direct (@function_name) or indirect (function pointer)
-                match callee {
-                    CallOpCallable::Direct(identifier) => {
-                        let name = identifier.to_string();
-                        // LLVM intrinsics use dots in IR; Pliron IR identifiers use underscores.
-                        let fixed_name = if name.starts_with("llvm_") {
-                            name.replace('_', ".")
-                        } else {
-                            // Strip cuda_oxide_device_ prefix from call targets to match
-                            // the stripped function definitions (clean export names).
-                            strip_device_prefix(&name)
-                        };
-                        is_convergent_call = Self::is_convergent_intrinsic(&fixed_name);
-                        write!(output, " @{fixed_name}(").unwrap();
-                    }
-                    CallOpCallable::Indirect(val) => {
-                        write!(output, " ").unwrap();
-                        self.export_value(val, value_names, output).unwrap();
-                        write!(output, "(").unwrap();
-                    }
-                }
-
-                // Export call arguments with their types
-                for (i, arg) in op_ref.operands().enumerate() {
-                    if i > 0 {
-                        write!(output, ", ").unwrap();
-                    }
-                    self.export_type(arg.get_type(self.ctx), output)?;
-                    write!(output, " ").unwrap();
-                    self.export_value(arg, value_names, output)?;
-                }
-
-                // Add convergent attribute reference for sync intrinsics
-                if is_convergent_call {
-                    writeln!(output, ") #0").unwrap();
-                    self.convergent_used = true;
-                } else {
-                    writeln!(output, ")").unwrap();
-                }
+                self.export_call_op(&op_ref, call, value_names, output)?;
             }
 
             // --- Inline Assembly ---
             id if id == ops::InlineAsmOp::get_opid_static() => {
                 let inline_asm = op_obj.as_ref().downcast_ref::<ops::InlineAsmOp>().unwrap();
-                let asm_template = inline_asm.asm_template(self.ctx);
-                let constraints = inline_asm.constraints(self.ctx);
-                let is_convergent = inline_asm.is_convergent(self.ctx);
-
-                // Check if there's a result
-                let has_result = op_ref.get_num_results() > 0;
-
-                if has_result {
-                    let res = op_ref.get_result(0);
-                    let res_name = value_names.get(&res).unwrap();
-                    let res_ty = res.get_type(self.ctx);
-                    write!(output, "  {res_name} = call ").unwrap();
-                    self.export_type(res_ty, output)?;
-                } else {
-                    write!(output, "  call void").unwrap();
-                }
-
-                // Format: call <type> asm sideeffect "<template>", "<constraints>"(<args>...)
-                write!(
-                    output,
-                    " asm sideeffect \"{asm_template}\", \"{constraints}\"("
-                )
-                .unwrap();
-
-                // Export input operands with types
-                for (i, arg) in op_ref.operands().enumerate() {
-                    if i > 0 {
-                        write!(output, ", ").unwrap();
-                    }
-                    self.export_type(arg.get_type(self.ctx), output)?;
-                    write!(output, " ").unwrap();
-                    self.export_value(arg, value_names, output)?;
-                }
-
-                // Add convergent attribute reference if needed
-                if is_convergent {
-                    writeln!(output, ") #0").unwrap();
-                    self.convergent_used = true;
-                } else {
-                    writeln!(output, ")").unwrap();
-                }
+                self.export_inline_asm_op(&op_ref, inline_asm, value_names, output)?;
             }
 
             // --- Multi-Output Inline Assembly ---
@@ -310,92 +191,7 @@ impl<'a> ModuleExportState<'a> {
                     .as_ref()
                     .downcast_ref::<ops::InlineAsmMultiOp>()
                     .unwrap();
-                let asm_template = inline_asm.asm_template(self.ctx);
-                let constraints = inline_asm.constraints(self.ctx);
-                let is_convergent = inline_asm.is_convergent(self.ctx);
-                let num_results = op_ref.get_num_results();
-
-                if num_results == 0 {
-                    // Void return - simple case
-                    write!(output, "  call void").unwrap();
-                    write!(
-                        output,
-                        " asm sideeffect \"{asm_template}\", \"{constraints}\"("
-                    )
-                    .unwrap();
-
-                    for (i, arg) in op_ref.operands().enumerate() {
-                        if i > 0 {
-                            write!(output, ", ").unwrap();
-                        }
-                        self.export_type(arg.get_type(self.ctx), output)?;
-                        write!(output, " ").unwrap();
-                        self.export_value(arg, value_names, output)?;
-                    }
-
-                    if is_convergent {
-                        writeln!(output, ") #0").unwrap();
-                        self.convergent_used = true;
-                    } else {
-                        writeln!(output, ")").unwrap();
-                    }
-                } else {
-                    // Multi-output: returns a struct, need extractvalue for each
-                    // Step 1: Build the struct type string
-                    let mut struct_type = String::from("{");
-                    for i in 0..num_results {
-                        if i > 0 {
-                            struct_type.push_str(", ");
-                        }
-                        let res_ty = op_ref.get_result(i).get_type(self.ctx);
-                        let mut ty_str = String::new();
-                        self.export_type(res_ty, &mut ty_str)?;
-                        struct_type.push_str(&ty_str);
-                    }
-                    struct_type.push('}');
-
-                    // Step 2: Generate the asm call returning the struct
-                    // We need a temporary name for the struct result
-                    // Use the first result's name with "_struct" suffix
-                    let first_res = op_ref.get_result(0);
-                    let first_res_name = value_names.get(&first_res).unwrap();
-                    let struct_result_name = format!("{first_res_name}_struct");
-
-                    write!(output, "  {struct_result_name} = call {struct_type}").unwrap();
-                    write!(
-                        output,
-                        " asm sideeffect \"{asm_template}\", \"{constraints}\"("
-                    )
-                    .unwrap();
-
-                    for (i, arg) in op_ref.operands().enumerate() {
-                        if i > 0 {
-                            write!(output, ", ").unwrap();
-                        }
-                        self.export_type(arg.get_type(self.ctx), output)?;
-                        write!(output, " ").unwrap();
-                        self.export_value(arg, value_names, output)?;
-                    }
-
-                    if is_convergent {
-                        writeln!(output, ") #0").unwrap();
-                        self.convergent_used = true;
-                    } else {
-                        writeln!(output, ")").unwrap();
-                    }
-
-                    // Step 3: Generate extractvalue for each result
-                    for i in 0..num_results {
-                        let res = op_ref.get_result(i);
-                        let res_name = value_names.get(&res).unwrap();
-
-                        writeln!(
-                            output,
-                            "  {res_name} = extractvalue {struct_type} {struct_result_name}, {i}"
-                        )
-                        .unwrap();
-                    }
-                }
+                self.export_inline_asm_multi_op(&op_ref, inline_asm, value_names, output)?;
             }
 
             // --- Casts ---
@@ -474,51 +270,6 @@ impl<'a> ModuleExportState<'a> {
             }
         }
 
-        Ok(())
-    }
-
-    pub(super) fn export_binop(
-        &self,
-        op_name: &str,
-        op: Ptr<Operation>,
-        value_names: &HashMap<Value, String>,
-        output: &mut String,
-    ) -> Result<(), String> {
-        let op_ref = op.deref(self.ctx);
-        let res = op_ref.get_result(0);
-        let lhs = op_ref.get_operand(0);
-        let rhs = op_ref.get_operand(1);
-        let res_name = value_names.get(&res).unwrap();
-
-        write!(output, "  {res_name} = {op_name} ").unwrap();
-        self.export_type(lhs.get_type(self.ctx), output)?;
-        write!(output, " ").unwrap();
-        self.export_value(lhs, value_names, output)?;
-        write!(output, ", ").unwrap();
-        self.export_value(rhs, value_names, output)?;
-        writeln!(output).unwrap();
-        Ok(())
-    }
-
-    pub(super) fn export_cast(
-        &self,
-        op_name: &str,
-        op: Ptr<Operation>,
-        value_names: &HashMap<Value, String>,
-        output: &mut String,
-    ) -> Result<(), String> {
-        let op_ref = op.deref(self.ctx);
-        let res = op_ref.get_result(0);
-        let val = op_ref.get_operand(0);
-        let res_name = value_names.get(&res).unwrap();
-
-        write!(output, "  {res_name} = {op_name} ").unwrap();
-        self.export_type(val.get_type(self.ctx), output)?;
-        write!(output, " ").unwrap();
-        self.export_value(val, value_names, output)?;
-        write!(output, " to ").unwrap();
-        self.export_type(res.get_type(self.ctx), output)?;
-        writeln!(output).unwrap();
         Ok(())
     }
 }
